setup vscode-test
