+++ .vscode-test.js
@@ -1,3 +1,3 @@
 const { defineConfig } = require("@vscode/test-cli");
 
-module.exports = defineConfig({ files: "dist/test/**/*.test.js" });
+module.exports = defineConfig({ files: "test/**/*.test.js" });
