+++ package.json
@@ -16,21 +16,14 @@
     "email": "yaki-shake@up-tri.me"
   },
   "homepage": "https://github.com/up-tri/oud-intellisense",
-  "categories": [
-    "Other"
-  ],
-  "activationEvents": [
-    "onLanguage:oudia"
-  ],
+  "categories": ["Other"],
+  "activationEvents": ["onLanguage:oudia"],
   "main": "./dist/extension.js",
   "contributes": {
     "languages": [
       {
         "id": "oudia",
-        "extensions": [
-          ".oud",
-          ".oud2"
-        ],
+        "extensions": [".oud", ".oud2"],
         "configuration": "./language-configuration.json"
       }
     ],
@@ -49,23 +42,26 @@
   },
   "scripts": {
     "vscode:prepublish": "tsc",
-    "build": "tsc --project tsconfig.build.json",
-    "build:watch": "tsc --project tsconfig.build.json --watch",
+    "build": "tsc --project tsconfig.json",
+    "build:test": "tsc --project tsconfig.test.json",
+    "build:watch": "tsc --project tsconfig.json --watch",
     "lint": "biome format .",
     "lint:fix": "biome format --write .",
-    "test": "vscode-test"
+    "test": "node_modules/.bin/vscode-test"
   },
   "devDependencies": {
     "@babel/core": "^7.25.2",
     "@babel/preset-env": "^7.25.4",
     "@babel/preset-typescript": "^7.24.7",
     "@biomejs/biome": "1.9.0",
+    "@types/mocha": "10.0.10",
     "@types/node": "^20.0",
     "@types/vscode": "1.93.0",
     "@vscode/test-cli": "0.0.10",
     "@vscode/test-electron": "2.4.1",
     "babel-jest": "^29.7.0",
     "glob": "^11.0.0",
+    "mocha": "10.8.2",
     "typescript": "~5.5.4"
   },
   "dependencies": {
