+++ test/resources/mock/eki_patterns.oud
