+++ test/resources/mock/empty_ressya_patterns.oud
