+++ test/resources/mock/ressya_patterns.oud
