+++ test/resources/mock/too_deep_file_pattern.oud
