--- tsconfig.build.json
@@ -1,12 +1,0 @@
-{
-  "compilerOptions": {
-    "module": "commonjs",
-    "target": "es6",
-    "outDir": "dist",
-    "lib": ["es6"],
-    "sourceMap": true,
-    "strict": true
-  },
-  "include": ["src/**/*.ts"],
-  "exclude": ["node_modules", ".vscode-test"]
-}
