+++ tsconfig.json
@@ -1,13 +1,23 @@
 {
-  "extends": "./tsconfig.build.json",
   "compilerOptions": {
     "types": [
-      "mocha",
       "node"
     ],
+    "lib": [
+      "es6"
+    ],
+    "module": "commonjs",
+    "target": "es6",
+    "outDir": "dist",
+    "sourceMap": false,
+    "strict": true
   },
   "include": [
     "src/**/*.ts",
+  ],
+  "exclude": [
+    "node_modules",
+    ".vscode-test",
     "test/**/*.ts"
   ]
 }
